Admin pages

+++ database/migrations/2022_05_24_085946_create_daily_coins_table.php
@@ -21,7 +21,7 @@ class CreateDailyCoinsTable extends Migration
             $table->double('market_cap')->nullable();
             $table->double('volume_24h')->nullable();
             $table->bigInteger('supply')->nullable();
-            $table->integer('total_stakes')->nullable();
+            $table->bigInteger('total_stakes')->nullable();
             $table->integer('staked_nodes')->nullable();
             $table->timestamps();
         });
